cometd_app/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
avro==1.11.3
bitarray==2.9.2
bitstring==4.1.4
certifi==2024.2.2
charset-normalizer==3.3.2
confluent-kafka==2.3.0
frozenlist==1.4.1
grpcio==1.62.1
grpcio-tools==1.62.1
idna==3.6
multidict==6.0.5
protobuf==4.25.3
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
yarl==1.9.4
